Refactor UI helpers into dedicated utility modules
