+++ src/components/Layout.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from "react";
 import { Outlet, useNavigate } from "react-router-dom";
 import Navigation from "@/components/ui/navigation";
-import { useAuth } from "@/hooks/useAuth";
+import { useAuth } from "@/hooks/use-auth-context";
 
 const Layout = () => {
   const navigate = useNavigate();
