+++ src/hooks/useUserRole.tsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react';
-import { useAuth } from './useAuth';
+import { useAuth } from './use-auth-context';
 import { supabase } from '@/integrations/supabase/client';
 
 export type UserRole = 'admin' | 'moderator' | 'user';
