+++ src/pages/EquipmentStore.tsx
@@ -20,7 +20,7 @@ import {
 } from "lucide-react";
 import { useToast } from "@/hooks/use-toast";
 import { supabase } from "@/integrations/supabase/client";
-import { useAuth } from "@/hooks/useAuth";
+import { useAuth } from "@/hooks/use-auth-context";
 import { useGameData, type PlayerSkills } from "@/hooks/useGameData";
 
 interface EquipmentItem {
