+++ src/pages/StreamingPlatforms.tsx
@@ -25,7 +25,7 @@ import {
   TrendingUp,
 } from "lucide-react";
 import { supabase } from "@/integrations/supabase/client";
-import { useAuth } from "@/hooks/useAuth";
+import { useAuth } from "@/hooks/use-auth-context";
 import { useGameData } from "@/hooks/useGameData";
 import type { Database } from "@/integrations/supabase/types";
